requirements.txt
diffusers
pandas
unicode-slugify